Versión 0.5.7
- Se completa la funcionalidad 5 como se tiene en el diagrama de interacción.

+++ src/gestionAplicacion/usuario/TarjetaCinemar.java
@@ -52,7 +52,7 @@ public class TarjetaCinemar implements IBuyable {
 		return dueno;
 	}
 
-	public void setSaldo(int saldo) {
+	public void setSaldo(double saldo) {
 		this.saldo = saldo;
 	}
